+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "Yearnfarmer",
+  "name": "DAOVaultMediumUSDT",
   "license": "UNLICENSED",
   "scripts": {
     "codegen": "graph codegen",
